+++ src/types.h
@@ -132,6 +132,7 @@ struct type_s {
 
 #define Type(typetag, ...) new (type_t, .tag = typetag, .__data.typetag = {__VA_ARGS__})
 #define INT_TYPE Type(BigIntType)
+#define INT64_TYPE Type(IntType, .bits = TYPE_IBITS64)
 #define NUM_TYPE Type(NumType, .bits = TYPE_NBITS64)
 #define NewFunctionType(ret, ...)                                                                                      \
     _make_function_type(ret, sizeof((arg_t[]){__VA_ARGS__}) / sizeof(arg_t), (arg_t[]){__VA_ARGS__})